requirements.txt
torch~=2.4.0+cu121
torchvision~=0.19.0+cu121
numpy~=1.26.4
pillow~=10.4.0
pyyaml
matplotlib~=3.9.2
tqdm~=4.66.5
sklearn
scikit-learn~=1.5.1
opencv-python~=4.11.0.86
albumentations~=2.0.8
unet~=0.8.0
pandas~=2.2.2
tensorflow~=2.18.0
scikit-image~=0.24.0
pydensecrf~=1.0rc3